New console implemented.
